add bti main file

+++ src/bti.py
@@ -1,12 +1,13 @@
 import nltk
 import os
+import datetime
 import csv
 
 from striprtf.striprtf import rtf_to_text
 from nltk.tokenize import sent_tokenize
-
-
 nltk.download("punkt")
+
+VERSION = "0.0.1"
 
 def get_rtf_as_sentences(file_path):
     sentences = []
@@ -45,22 +46,24 @@ def create_data_structure(sentences, country, year):
         line_counter += 1
     return rows
 
-def convert_to_csv(file_path, output_dir="./data/bti/raw-csv/"):
+def convert_to_csv(file_path, output_dir="./data/bti/raw-csv/", overwrite=False):
+    current_datetime_utc = datetime.datetime.utcnow()
+    # convert the datetime object to ISO format with 'Z' indicating UTC timezone
+    current_datetime_utc_iso = current_datetime_utc.replace(microsecond=0).isoformat() + 'Z'
     file_name = os.path.splitext(os.path.basename(file_path))[0]
     c_n_y = file_name.split("_")
     country = c_n_y[0].lower()
     year = c_n_y[1].replace("BTI", "")
     output_file = os.path.join(output_dir, file_name + ".csv")
-    sentences = get_rtf_as_sentences(path)
-    rows = create_data_structure(sentences, country, year)
-    with open(output_file, mode="w", newline="") as csv_file:
-        writer = csv.writer(csv_file)
-        for row in rows:
-            writer.writerow(row)
-
-
-
-# test
-path = "./data/bti/raw-rtf/Moldova_BTI2003.rtf"
-convert_to_csv(path)
-
+    if os.path.exists(output_file) and overwrite is False:
+        return (False, os.path.basename(file_path), current_datetime_utc_iso, VERSION)
+    else:
+        sentences = get_rtf_as_sentences(file_path)
+        rows = create_data_structure(sentences, country, year)
+        with open(output_file, mode="w", newline="") as csv_file:
+            writer = csv.writer(csv_file)
+            header = ("sentence", "section", "country", "year", "source")
+            writer.writerow(header)
+            for row in rows:
+                writer.writerow(row)
+        return (True, os.path.basename(file_path), current_datetime_utc_iso, VERSION)
